// src/helpers/node-order-type.ts
export enum NodeOrderType
{
    none = "none",
    orderByType = "orderByType",
    orderByTypeByName = "orderByTypeByName",
    orderByTypeByAccessor = "orderByTypeByAccessor",
    orderByTypeByAccessorByName = "orderByTypeByAccessorByName",
    groupByType = "groupByType",
    groupByTypeOrderByName = "groupByTypeOrderByName",
    groupByTypeOrderByAccessor = "groupByTypeOrderByAccessor",
    groupByTypeOrderByAccessorByName = "groupByTypeOrderByAccessorByName",
    groupByTypeByAccessorOrderByTypeByAccessorByName = "groupByTypeByAccessorOrderByTypeByAccessorByName"
}